backend/authentication/serializers.py
from rest_framework import serializers
from .models import User
from profiles.models import StudentProfile, TPOProfile, CompanyProfile
from django.contrib.auth.hashers import make_password
from django.contrib.auth import authenticate
from rest_framework_simplejwt.tokens import RefreshToken

class RegistrationSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['email', 'password', 'role']

    def create(self, validated_data):
        validated_data['password'] = make_password(validated_data['password'])
        user = super().create(validated_data)

        if user.role == 'student':  
            StudentProfile.objects.create(user=user)
        elif user.role == 'tpo':
            TPOProfile.objects.create(user=user)
        elif user.role == 'company':
            CompanyProfile.objects.create(user=user)

        return user    

    
class LoginSerializer(serializers.Serializer):
    email = serializers.EmailField()
    password = serializers.CharField()

    def validate(self, data):
        email = data.get('email')
        password = data.get('password')
        user = authenticate(email=email, password=password)
        
        if not user:
            raise serializers.ValidationError("Invalid email or password")
        
        refresh = RefreshToken.for_user(user)
        return {
            'refresh': str(refresh),
            'access': str(refresh.access_token),
            'user': {
                'id': user.id,
                'email': user.email,
                'role': user.role
            }
        }