backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.7.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
joblib==1.4.2
numpy==2.1.3
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
scikit-learn==1.5.2
scipy==1.14.1
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.13.2
